==> notebooks/02_cleaning_fact_events.ipynb <==
### Cleaning Overview
- Input: Raw GA4 event-level Parquet snapshot
- Objective: Prepare analytics-ready dataset
- Key cleaning steps:
  - Null handling
  - Deduplication
  - Type standardization
  - Feature derivation
- Output: Clean Parquet for downstream analysis


# importing necessary libraries

import pandas as pd
from pathlib import Path
import numpy as np

# since sample file is included in repo, adding toggle to switch file based on scenario

use_sample = False

if use_sample == True:
    RAW_DATA_PATH = Path("../data/sample")
    raw_file = RAW_DATA_PATH / "fact_events_raw_sample.parquet"
else:
    RAW_DATA_PATH = Path("../data/raw")
    raw_file = RAW_DATA_PATH / "fact_events_raw.parquet"

CLEAN_DATA_PATH = Path("../data/cleaned")


df = pd.read_parquet(raw_file)
df.info(show_counts=True)

# handling nulls and duplicates

df['engagement_time_msec'] = df['engagement_time_msec'].fillna(0)
df = df.drop_duplicates(subset= ['user_id','event_name','event_ts'])

# normalizing columns

for col in ['traffic_smedium', 'traffic_source', 'traffic_sname']:
    df[col] = df[col].fillna('').str.lower().str.strip()


# derriving categories for marketing channel

conditions = [
    # privacy
    (df['traffic_smedium'] == '(data deleted)') |
    (df['traffic_source'] == '(data deleted)' ) |
    (df['traffic_sname'] == '(data deleted)'),
    
    # direct traffic
    (df['traffic_smedium'] == '(none)')         &
    (df['traffic_source'] == '(direct)'),

    # organic traffic
    (df['traffic_smedium'] == 'organic'),

    #paid
    (df['traffic_smedium'] == 'cpc'),

    #referral
    (df['traffic_smedium'] == 'referral')
]

choices = [
    'unknown_privacy',
    'direct',
    'organic',
    'paid',
    'referral'
]

# adding marketing channel category

df['marketing_channel'] = np.select(conditions,choices,default='others')


# categorzing columns and adding flags

category_columns = ['event_name','device','country','marketing_channel']
for col in category_columns:
    df[col] = df[col].astype('category')

df['is_purchase'] = df['transaction_id'].notna()
df['is_engaged_event'] = df['engagement_time_msec'] > 0

# exporting cleaned parquet file to cleaned

OUTPUT_PATH = CLEAN_DATA_PATH / "fact_events_cleaned.parquet"

df.to_parquet(OUTPUT_PATH,index = False)
